requirements.txt
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
Click==7.0
cryptography==2.8
Flask==1.1.1
future==0.18.2
gunicorn==20.0.4
idna==2.8
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
pycparser==2.19
python-telegram-bot==12.3.0
requests==2.22.0
six==1.14.0
tornado==6.0.3
urllib3==1.25.8
Werkzeug==0.16.1
